Support both timer modes and value retrieval
The interrupt may may follow the alarm or timer flag value
or if not in alarm, be pulsed every timer priod.  Enable making
this choice when setting the timer.  Note, the timer flag must
still be cleared manually (it does not pulse).

It's useful to be able to inspect the timer settings
and current value.  However retrieving the value is tricky
since there is no buffering while the timer is running.
Multiple retrieval attempts are used to verify consistency
before the value is returned.

Fixed bug where status/timer_control values could get
out of sync with interface.

Updated Ketwords & fixed missing zeroClock.

Signed-off-by: Chris Evich <chris-arduino@anonomail.me>