Add clearVoltLow() member function
It doesn't matter if you set the VL_Seconds bit low or
high, it is always cleared when you write to it.  Added
a member function that only clears the bit, retaining
current date/time values.  Made setTime() to mask off
the VL_Seconds bit, so it's always cleared no matter what.

Signed-off-by: Chris Evich <chris-arduino@anonomail.me>