fix days-in-year calculation bug.
The algorithm idea was accurate but the lookup table was wrong.
It's also a bit simpler to just include the 0th element since
the conditionals to test for it likely occupy more than one
byte anyway.  Tested this version with multiple months in and out
of leap-years and it's producing correct results now.

Signed-off-by: Chris Evich <chris-arduino@anonomail.me>