Rtc_Pcf8563 specific version of leapYearsBetween
In order to calculate a duration which could span
any number of leap and non-leap years, it's necessary
to account for them.  This method returns the number
of leap-days for years from start to end (exclusive).
The century-bit handling behavior matches that of
the chip and other calculation methods here.

Signed-off-by: Chris Evich <chris-arduino@anonomail.me>